requirements.txt
fastapi
uvicorn
pydantic
google-generativeai
pillow
nudenet